Move engines code to helpers
